requirements.txt
psycopg2-binary
pandas
yfinance
python-dotenv
